Smoothed out obvious wrinkles with compiler and fixed inconsistencies.

+++ src/scstruct.h
@@ -49,9 +49,9 @@ struct SCRIPT {
 struct SCRIPT_COMPILE_STATE {
    uint8_t flags;
    char token_iter[SCRIPT_TOKEN_ITER_SZ_MAX + 1];
-   size_t token_iter_sz;
+   uint16_t token_iter_sz;
    struct SCRIPT_STEP* steps;
-   size_t steps_sz;
+   uint16_t steps_sz;
    int16_t last_start;
 };
